Add rate limiting to auth endpoints and enforce whole-number shares
- Rate limit register/login to 10 requests per IP per minute (429 on excess)
- Change TradeRequest.shares from float to int to prevent fractional
  share rounding issues

+++ backend/app/schemas/__init__.py
@@ -97,7 +97,7 @@ class TradeRequest(BaseModel):
     season_id: str
     stock_symbol: str = Field(..., max_length=10)
     transaction_type: str = Field(..., pattern="^(BUY|SELL)$")
-    shares: float = Field(..., gt=0)
+    shares: int = Field(..., gt=0)
 
 
 class TradeResponse(BaseModel):
